chore: bundle extension
