+++ editors/vscode/.vscodeignore
@@ -2,7 +2,7 @@
 !icon.png
 !LICENSE
 !node_modules/
-!out/extension.js
+!out/main.js
 !package-lock.json
 !package.json
 !server/rome_lsp
